test(governance): extend GovernancePayload test with starknet types
